SST_PLATFORM_STORAGE: Corrected Content Set Definitions.
There were a number of packages that were defined for multiple architectures
when they were only destined for one particular architecture.

  libblockdev-s390 - s390x only
  libpmemobj++ - x86_64 only
  libpmemobj++-devel - x86_64 only
  pmdk-convert - x86_64 only

There was a set of packages that we decided was no longer needed to be included.

  libvmem*
  libvmalloc*

There was also a package that was identified as only being available in RHEL
(not built in Fedora), so it is now listed as a placeholder.
  pmreorder

There were some packages that were also erron