sst_networking: add package placeholder for "NetworkManager-config-connectivity-redhat"
The "NetworkManager-config-connectivity-redhat" package is a subpackage
of the "NetworkManager" source RPM.

On RHEL we build this package, but on Fedora there is instead
"NetworkManager-config-connectivity-fedora". This causes the workload to
be invalid.

Fix that by adding it as "package_placeholders".

Note that since the package is a subpackage, it has essentially the same
build requirements as "NetworkManager".