sst_cs_software_mgmt: add yum and yum-utils
Upstream still has these two packages separate, so we need to keep them
as such in RHEL-9, too.

They're both pretty small packages whose only purpose is to ship various
YUM-3 compatibility CLI shims and man page symlinks.

It is possible to reconsider this in the future and perhaps merge the
shims into the mainline dnf/dnf-utils packages (and make "yum" and
"yum-utils" into Provides:), but that would require a bit more thought
and would need to happen in upstream first.

Also note that dnf-utils is removed here - it is just a Provides: of the
yum-utils package in upstream right now.