Use rubygem-tilt placeholder to reduce Ruby dependency chain.
Tilt has gazillion dependencies and what worser, half of them are
circular. Luckily, they test just possible runtime dependencies which
will be unused at the end. They are enabled on Fedora, just because they
can be enabled. Therefore cut them off to reduce the Ruby dependency
chain.

Also note that when dependency is not available, the tests are just
skipped.