Add unwanted python-networkx dependencies
A lot of unnecessary dependencies were originally used as BR and as weak
dependencies. As the optional functionality provided by these packages
is not needed for setools, we decided to drop them from BR and leave
them as weak dependencies only, see
https://src.fedoraproject.org/rpms/python-networkx/pull-request/7

Also we want to be sure that these components are not assigned to
sst_security_selinux and therefore mark them as unwanted even though
Recommends: should be ignored.

Signed-off-by: Petr Lautrbach <plautrba@redhat.com>