sst_cs_software_mgmt: bring Lua CLI back
We decided to keep the "lua" binary package (the CLI interpreter) around
after all.  It's built from the same Lua SRPM that we already own (as an
embedded interpreter in RPM), so there's not much additional maintenance
burden associated with the CLI interpreter itself (which alone can be
quite useful).