Omit python-hs-dbus-signature from supported libraries in config file
The library is purely for testing. Where a Stratis Python library
previously unconditionally relied on this package for testing, tests have
been rewritten so that it is possible to run the subset of tests that do
not require this package without loading the packages. Only these tests,
that do not require python-hs-dbus-signature, are used in upstream Fedora
packaging. Therefore this package does not need to be specified at all
in the configuration file.