Add missing binary packages created by linux-firmware component
We were listing only linux-firmware, but there are other packages
created by linux-firmware src.rpm that we were not listing, and
that are installed by default or there is at the moment a kernel module
which is built and requires the firmware provided by them.

There is one exception here though, I'm not adding libertas-sd8787-firmware,
because it seems we probably should not be building on centos/rhel the kernel
modules that require it anymore (mwifiex_sdio, btmrvl_sdio), so leaving this
out for now. And on errata this specific firmware package also falls
into the "not in Product Listings" category already.

See also https://issues.redhat.com/browse/RHELPLAN-46582

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>