Update the automotive environment and workloads
This adds a bunch of packages based on new requirements from an
automotive vendor.  This generally covers C++, compression and
multimedia.

Development subpackages were added for requested libraries to the
off-vehicle workload.

Some base libraries are required explicitly to denote their nature of
providing top-level APIs in the automotive context.

Signed-off-by: Petr Šabata <contyk@redhat.com>