automotive-repositories: switch to production c9s composes
Switches the CS9 repositories to the production compose as the
development compose is unstable. Currently the annobin packages in the
development compose have bugs that block it from reaching the production
compose and is a case where we would be introducing a buggy state by
using the development compose as an input.

Also drops buildroot and epel as we don't want to introduce non c9s
content into the AutoSD distribution by default.

Also adds the autosd CBS packages and shift that plus the Automotive CBS
packages to a higher priority layer (would make sense that what they fork
take priority over C9S).