Remove automotive-sig from automotive-repositories
The AutoSD project is using the content-resolver output to define the
product definition (contents) of the AutoSD repositories (aka
product-build). This commit strips the automotive-sig repositories and
packages from the content-resolver workload/repositories because they're
not considered part of the AutoSD product and are dirtying the product
definition (and muddying the line between automotive-sig and AutoSD).

This is needed to create a cleaner divide between the AutoSD project and
the automotive-sig project. Later if these packages are needed in
AutoSD, they need to be properly promoted to the autosd tag in order to
add them here.