Update the accessibility workload
Together with people from the Red Hat's Linux accessibility group, we've
agreed on only shipping the espeak-ng speech synthesizer (owned by
sst_cs_infra_services) in RHEL 10 and drop flite and festival - both are
no longer maintained upstream for some time and they do have their set of
problems. espeak-ng is actively maintained upstream and has a brighter
future in RHEL 10. The only downside of this plan is that espeak-ng sounds
"robotic" in opposed to festival that is using the natural voice. If there
will be demands in the future for the natural voice synthesizer, then we
might look into how RHVoice is doing (currently not even in Fedora). See
https://issues.redhat.com/browse/DESKTOP-660 for the internal
discussion.

Drop at-spi2-atk as it has been merged into at-spi2-core.

Explicitly add liblouis which is a braille translator.