Drop Qt 5 from ELN and introduce Qt 6 there
We are only supporting one Qt version per major release of RHEL and as
Qt 5 is on the end of its development cycle, we are removing it from
RHEL 10 and bringing Qt 6 there which is the current stable Qt release
in development that will see more of the Wayland specific features
developed there. See https://issues.redhat.com/browse/DESKTOP-734 for
internal ticket.