Drop evolution-mapi from RHEL 10
evolution-mapi, which is a connector to a Microsoft Exchange server
using a MAPI protocol. This protocol had been deprecated since 2007 or
so, disabled by default on the Exchange 20123 server, +/-. The connector
is currently used only for very old servers, like Exchange 2003, 2007
and 2010. Microsoft aims to an EWS protocol since 2007, while it does
not receive any new features now, as they develop a new protocol, called
Microsoft Graph API, which is only for the Exchange Online, not for
On-Premise Exchange servers (thus the EWS is still the right thing to
use), at least as of now.

By removing the evolution-mapi, its dependency openchange package can be
removed as well. The OpenChange project is dead upstream for years. The
openchange package depends on the samba package, thus whenever samba is
updated, both openchange and evolution-mapi need to be rebuild. By
removing these two it'll make life easier for the samba folks as well
(it happened in the past that the rebuild had been forgotten).

See https://issues.redhat.com/browse/DESKTOP-692 (internal to RH) for
more information.