Remove pigz
pigz was used to compress and decompress gzip'd kernel modules.  It's no longer
used AFAICT anywhere in anaconda or for kernel module compression.  It can
be dropped as a dependency, which should eventually lead to its removal
from RHEL and Fedora.  It is also worth noting that pigz hasn't been
updated in YEARS and also points to the packages' lack of use.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>