Move Thunderbird and Firefox from ELN to ELN Extras
We're using different packaging in RHEL (and in RHEL 10 those will
be shipped as Flatpaks only) than in Fedora so the dependencies are
not the same and the ELN is being "polluted" by packages that won't
be needed in RHEL at all. On the other hand we're still interested in
knowing how Thundebird and Firefox comply with all the ELN specific
changes (extra hardening flags, ..) so ELN Extras is better place
for these.