Remove anaconda-user-help from RHEL
The main dependency for build-in help feature in Anaconda is removed
from RHEL, so we are dropping this functionality.

It was removed by commit 83dd9bceec315ac0cddc02df4f4b326ff6fe6ab5 which
is main dependency for Yelp application. Even if this is resolved by
using Firefox as replacement, the drawback of required space and
resources is too big for Anaconda to overcome.

Based on our research it seems that this functionality is not critical
nor much utilized so we are going ahead to drop this.