Make scap-security-guide a placeholder
scap-security-guide in Fedora includes all OS data, but in RHEL it only
includes data for that RHEL version.  Since that RHEL version specific
data needs to be created as part of development, it cannot be so built
in ELN or imported when RHEL branches.  Also, there is a RHEL-specific
rule-playbooks subpackage.  As such, scap-security-guide fits the
criteria of a placeholder, and handling it as such avoids issues.