Remove nvme-cli from sst_arch_hw_eln.yaml
The nvme-cli program should no longer be pulled in by arch_hw.  We used to have an engineer on staff for nvme support, but those activities have been moved over to storage.  The nvme-cli will still be pulled in as a requirement for the anaconda workload.  This will make no functional change to the package set, it will simply clean up the how/why of what packages are included.