Drop zram-generator-defaults from base
Enabling swap-on-zram must be explicit action carried out by the admin.
To prevent accidental enablement let's drop the package which ships
default configuration. We have an example config in the main package so
enabling it is still very easy.

This change makes zram setup in ELN and C10s same as it is C9s.