Drop Qt5WebEngine dependent packages from Extras
While Qt5 is planned for EPEL 10, its WebEngine component is not.  cantor
and kaidan show no signs of being close to Qt6 conversion, so this disables
them in the meantime.  ghostwriter will move to Qt6 in 24.08, and marble's
usage is optional and can be disabled in spec.

FWIW Qt5WebKit is even more unwanted, but so far the only dependents are
kexi/kreport (which is about to be disabled in spec) and kdevelop (which is
also moving to Qt6 in 24.08).