Block dogtak-pki logging dependencies
The first step of bundling, which copies these dependencies from the
buildroot into the package, has already landed in Fedora.  The second step,
which is bundling them in the sources so as not to need standalone versions
even in the buildroot, has only landed in RHEL 10 so far; RHEL 11 will
certainly do the same, either in Fedora or ELN prior to branching, or in
RHEL during branching.