Stop skipping cbindgen
cbindgen is already a build dependency of mozjsNNN (usually bundled in RHEL
gjs post-branching), and could also be unbundled from firefox and
thunderbird. Its CLI only will soon be added to RHEL 9+ as a build
dependency of a new package, so it needs to be unblocked, while its -devel
packages are still needed in EPEL for building a few other packages.  For
now it will be built Fedora-style (with its rust-*-devel dependencies
blocked), but how to handle it long-term in ELN still needs to be decided.

https://github.com/fedora-eln/eln/issues/253