Drop dotnet8.0 from ELN
Microsoft has announced that .NET STS releases will now be supported for 24
months, meaning that it and the previous LTS will be supported until the
same date.  Therefore, for the purposes of ELN, there are no longer any
reasons not to simply focus on the latest stable version (whether it is STS
or LTS), as that is the best preparation for the next version to come.

https://devblogs.microsoft.com/dotnet/dotnet-sts-releases-supported-for-24-months/
https://dotnet.microsoft.com/en-us/platform/support/policy/dotnet-core