Require only dnf5 in ELN
While some packages still need to be ported from libdnf or python3-dnf and
will remain in ELN in the meantime, the goal is to fully transition to dnf5
for RHEL 11 and therefore dnf-4 should not be listed as required.

dnf5 also provides/obsoletes microdnf, so the minimal workload is obsolete.