Cleanup pcp-grafana unwanted packages
chan and HdrHistogram_c are used by pcp.  While that team does not want
their -devel packages shipped even in CRB -- and nothing else in Fedora
depends thereon for them to be a benefit to EPEL anyway -- they are still
needed and therefore are not unwanted (by the current definition thereof).